refactor: NetworkService 클래스 구현 및 OpenMarketAPI 클래스 코드 간소화
